pyproject.toml
 [build-system]
 requires = ["hatchling>=1.24.2"]
 build-backend = "hatchling.build"
 
 [project]
 name = "document-intelligence-refinery"
 version = "0.1.0"
 description = "Agentic pipeline for structure-preserving extraction, chunking, indexing, and provenance from enterprise documents."
 readme = "README.md"
 requires-python = ">=3.10"
 license = { text = "MIT" }
 authors = [{ name = "TRP1 Week 3" }]
 dependencies = [
   "pydantic>=2.7.0",
   "pydantic-settings>=2.2.1",
   "pyyaml>=6.0.1",
   "pdfplumber>=0.11.4",
   "pymupdf>=1.24.7",
   "langdetect>=1.0.9",
   "python-dotenv>=1.0.1",
   "httpx>=0.27.0",
   "tqdm>=4.66.4",
 ]
 
 [project.optional-dependencies]
 dev = [
   "pytest>=8.2.0",
   "pytest-cov>=5.0.0",
   "ruff>=0.5.0",
 ]
 
 [tool.hatch.build.targets.wheel]
 packages = ["src/refinery", "src/models", "src/agents", "src/strategies"]
 
 [tool.ruff]
 line-length = 100
 target-version = "py310"
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
 addopts = "-q"
